[skip ci] RNGP - Ignore the `legacyWarningsEnabled` property if provided
Summary:
We decided to change the warning model for LegacyArch/NewArch.

I'm currently removing the infra to read the `legacyWarningsEnabled` Gradle property if provided.

Warnings will be enabled by default for all Legacy Arch users in the new model. 

This change was never shipped in a numbered version, so that's not breaking.

Changelog:
[Internal] [Changed] -

Differential Revision: D73591315